+++ NEWS.md
@@ -1,3 +1,5 @@
+# dialr (development version)
+
 
 # dialr 0.2.1