+++ NEWS.md
@@ -6,6 +6,10 @@
 
 * Add geocoding functions `get_geocode()` and `get_timezone()` (#5).
 
+## Minor changes
+
+* Rename "One shot methods" vignette to remove redundant "dialr".
+
 # dialr 0.2.1
 
 ## Performance improvements